+++ src/VaultSharp/V1/SecretsEngines/Database/IDatabaseSecretsEngine.cs
@@ -24,5 +24,91 @@ namespace VaultSharp.V1.SecretsEngines.Database
         /// The secret with the <see cref="UsernamePasswordCredentials" /> as the data.
         /// </returns>
         Task<Secret<UsernamePasswordCredentials>> GetCredentialsAsync(string roleName, string mountPoint = SecretsEngineDefaultPaths.Database, string wrapTimeToLive = null);
+
+        /// <summary>
+        /// This endpoint creates or updates a static role definition. 
+        /// Static Roles are a 1-to-1 mapping of a Vault Role to a user in a database which are automatically 
+        /// rotated based on the configured rotation_period. 
+        /// Not all databases support Static Roles, please see the database-specific documentation.
+        /// </summary>
+        /// <param name="roleName"><para>[required]</para>
+        /// Specifies the name of the role to create. </param>
+        /// <param name="staticRole"><para>[required]</para>
+        /// Specifies the request options. </param>
+        /// <param name="mountPoint"><para>[optional]</para>
+        /// The mount point for the Database backend. Defaults to <see cref="SecretsEngineDefaultPaths.Database" />
+        /// Provide a value only if you have customized the mount point.</param>        
+        /// <returns>The task</returns>
+        Task CreateStaticRoleAsync(string roleName, StaticRole staticRole, string mountPoint = SecretsEngineDefaultPaths.Database);
+
+        /// <summary>
+        /// This endpoint queries the static role definition.
+        /// </summary>
+        /// <param name="roleName"><para>[required]</para>
+        /// Specifies the name of the role to read. </param>
+        /// <param name="mountPoint"><para>[optional]</para>
+        /// The mount point for the Database backend. Defaults to <see cref="SecretsEngineDefaultPaths.Database" />
+        /// Provide a value only if you have customized the mount point.</param>   
+        /// <param name="wrapTimeToLive">
+        /// <para>[optional]</para>
+        /// The TTL for the token and can be either an integer number of seconds or a string duration of seconds.
+        /// </param>
+        /// <returns>The role info.</returns>
+        Task<Secret<StaticRole>> ReadStaticRoleAsync(string roleName, string mountPoint = SecretsEngineDefaultPaths.Database, string wrapTimeToLive = null);
+
+        /// <summary>
+        /// This endpoint returns a list of available static roles. 
+        /// Only the role names are returned, not any values.
+        /// </summary>
+        /// <param name="mountPoint"><para>[optional]</para>
+        /// The mount point for the Database backend. Defaults to <see cref="SecretsEngineDefaultPaths.Database" />
+        /// Provide a value only if you have customized the mount point.</param>   
+        /// <param name="wrapTimeToLive">
+        /// <para>[optional]</para>
+        /// The TTL for the token and can be either an integer number of seconds or a string duration of seconds.
+        /// </param>
+        /// <returns>The role names.</returns>
+        Task<Secret<ListInfo>> ReadAllStaticRolesAsync(string mountPoint = SecretsEngineDefaultPaths.Database, string wrapTimeToLive = null);
+
+        /// <summary>
+        /// This endpoint deletes the static role definition and revokes the database user.
+        /// </summary>
+        /// <param name="roleName"><para>[required]</para>
+        /// Specifies the name of the role to delete. </param>
+        /// <param name="mountPoint"><para>[optional]</para>
+        /// The mount point for the Database backend. Defaults to <see cref="SecretsEngineDefaultPaths.Database" />
+        /// Provide a value only if you have customized the mount point.</param>   
+        /// <returns>The task.</returns>
+        Task DeleteStaticRoleAsync(string roleName, string mountPoint = SecretsEngineDefaultPaths.Database);
+
+        /// <summary>
+        /// Generates a new set of STATIC credentials based on the named role.
+        /// </summary>
+        /// <param name="roleName"><para>[required]</para>
+        /// Specifies the name of the static role to get credentials for.</param>
+        /// <param name="mountPoint"><para>[optional]</para>
+        /// The mount point for the Database backend. Defaults to <see cref="SecretsEngineDefaultPaths.Database" />
+        /// Provide a value only if you have customized the mount point.</param>
+        /// <param name="wrapTimeToLive">
+        /// <para>[optional]</para>
+        /// The TTL for the token and can be either an integer number of seconds or a string duration of seconds.
+        /// </param>
+        /// <returns>
+        /// The secret with the <see cref="StaticCredentials" /> as the data.
+        /// </returns>
+        Task<Secret<StaticCredentials>> GetStaticCredentialsAsync(string roleName, string mountPoint = SecretsEngineDefaultPaths.Database, string wrapTimeToLive = null);
+
+        /// <summary>
+        /// This endpoint is used to rotate the Static Role credentials stored for a given role name. 
+        /// While Static Roles are rotated automatically by Vault at configured rotation periods, 
+        /// users can use this endpoint to manually trigger a rotation to change the stored password and 
+        /// reset the TTL of the Static Role's password.
+        /// </summary>
+        /// <param name="roleName"><para>[required]</para>
+        /// Specifies the name of the static role to rotate credentials for.</param>
+        /// <param name="mountPoint"><para>[optional]</para>
+        /// The mount point for the Database backend. Defaults to <see cref="SecretsEngineDefaultPaths.Database" />
+        /// Provide a value only if you have customized the mount point.</param>
+        Task RotateStaticCredentialsAsync(string roleName, string mountPoint = SecretsEngineDefaultPaths.Database);
     }
 }